+++ listenbrainz/webserver/static/js/tests/__mocks__/year-in-music-data.json
